backend/requirements.txt
fastapi
uvicorn
python-multipart
groq
Pillow
numpy
scikit-learn
python-dotenv
